% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getComments}
\alias{getComments}
\title{Retrieve a corpus of all comments of the thread passed to getComments}
\usage{
getComments(url)
}
\arguments{
\item{url}{Link to desired Reddit comment thread}
}
\value{
A corpus of all the comments in the thread
}
\description{
Retrieve a corpus of all comments of the thread passed to getComments
}
\examples{
getWords("https://www.reddit.com/r/funny/comments/4mno3n/pure_chaos/")
}

