% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{negWordCloud}
\alias{negWordCloud}
\title{Retrieve a vector of all the words in all the comments of the thread passed to getWords.}
\usage{
negWordCloud(words)
}
\arguments{
\item{words}{A vector of all the words in the comment thread}
}
\value{
A pretty wordcloud
}
\description{
Retrieve a vector of all the words in all the comments of the thread passed to getWords.
}
\examples{
getWords(words_vector) 
}

