% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{negWordCloud}
\alias{negWordCloud}
\title{Retrieve a wordcloud of the most used negative words in all the comments of the thread passed to negWordCloud
Ensure you install tm.lexicon.GeneralInquirer for this by running:
install.packages("tm.lexicon.GeneralInquirer", repos="http://datacube.wu.ac.at", type="source")}
\usage{
negWordCloud(words)
}
\arguments{
\item{words}{A vector of all the words in the comment thread}
}
\value{
A pretty wordcloud
}
\description{
Retrieve a wordcloud of the most used negative words in all the comments of the thread passed to negWordCloud
Ensure you install tm.lexicon.GeneralInquirer for this by running:
install.packages("tm.lexicon.GeneralInquirer", repos="http://datacube.wu.ac.at", type="source")
}
\examples{
getWords(words_vector) 
}

